# openapi.yaml
openapi: '3.0.3'
info:
  title: TMS (Transport Management System) API
  description: |
    Multi-tenant Transport Management System with OTP-based authentication.
    
    ## Authentication
    - All protected endpoints require a JWT access token in the Authorization header
    - Format: `Authorization: Bearer <token>`
    - Access tokens expire after 30 minutes
    - Refresh tokens expire after 7 days
    
    ## Roles
    - **SUPER_ADMIN**: Full system access, can create tenants/customers
    - **TENANT_ADMIN**: Manage their own tenant
    - **TENANT_USER**: Basic tenant access
  version: '1.0.0'
  contact:
    name: TMS API Support
    email: support@tms.example.com
servers:
  - url: http://localhost:8080/api
    description: Local development server
  - url: https://api.tms.example.com/api
    description: Production server
...
paths:
  /auth/request-otp:
    post:
      tags:
        - Authentication
      summary: Request OTP for login
      description: |
        Generates and sends a 6-digit OTP to the user's email address.
        OTP is valid for 5 minutes. Rate limited to 3 requests per 15 minutes per email.
      operationId: requestOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
              properties:
                email:
                  type: string
                  format: email
                  example: admin@example.com
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              example:
                success: true
                message: OTP sent to your email
        '400':
          description: Invalid email format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many requests (rate limit exceeded)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: Too many OTP requests. Please try again later.
  /auth/verify-otp:
    post:
      tags:
        - Authentication
      summary: Verify OTP and login
      description: |
        Verifies the OTP and returns JWT access and refresh tokens.
        OTP must be used within 5 minutes and can only be used once.
      operationId: verifyOTP
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - otp
              properties:
                email:
                  type: string
                  format: email
                  example: admin@example.com
                otp:
                  type: string
                  pattern: '^\d{6}$'
                  example: '123456'
      responses:
        '200':
          description: OTP verified successfully, tokens returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Login successful
                  data:
                    type: object
                    properties:
                      authToken:
                        type: string
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      refreshToken:
                        type: string
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      userType:
                        type: string
                        enum: [SUPER_ADMIN, TENANT_ADMIN, TENANT_USER]
                        example: TENANT_ADMIN
        '401':
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid request format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: |
        Uses a valid refresh token to obtain a new access token and refresh token pair.
        The old tokens are invalidated.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refreshToken
              properties:
                refreshToken:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: New tokens generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Token refreshed
                  data:
                    type: object
                    properties:
                      authToken:
                        type: string
                      refreshToken:
                        type: string
        '401':
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/context:
    get:
      tags:
        - Authentication
      summary: Get current user context
      description: Returns detailed information about the currently authenticated user
      operationId: getUserContext
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User context retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized - Invalid or missing token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/verify-otp or /auth/refresh
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message description
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: user@example.com
        fullName:
          type: string
          example: John Doe
        role:
          type: string
          enum: [SUPER_ADMIN, TENANT_ADMIN, TENANT_USER]
          example: TENANT_ADMIN
        tenantId:
          type: string
          format: uuid
          nullable: true
          example: 660e8400-e29b-41d4-a716-446655440001
        phone:
          type: string
          example: '+1234567890'
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
          example: '2024-01-15T10:30:00Z'
        updatedAt:
          type: string
          format: date-time
          example: '2024-12-20T14:25:00Z'
...